+++ build.mjs
@@ -8,7 +8,7 @@ await esbuild.build({
   platform: 'node',
   format: 'cjs',
   sourcemap: 'linked',
-  treeShaking: true,
-  minify: true,
+  // treeShaking: true,
+  // minify: true,
   mainFields: ['module', 'main'],  //  needed for jsonc-parse until they fix https://github.com/microsoft/node-jsonc-parser/issues/57
 });