+++ server/src/tests/cases/compiler/thisObject1.file2.c
@@ -1,0 +1,3 @@
+
+public foo() { return 1; }
+protected bar() { return 1; }