+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Fix: [Diagnostics are duplicated after running the build task #137](https://github.com/jlchmura/lpc-language-server/issues/137)
 -   Fix: [LPCDoc comment should not be captured by #include #140](https://github.com/jlchmura/lpc-language-server/issues/140)
+-   Fix: [Incorrect signature resolution when function decl does not have doc comment #139](https://github.com/jlchmura/lpc-language-server/issues/139)  
 -   Added `.h` to the file extensions that will activate the language server.
 
 ## 1.1.17